projectconfig/requirements.txt
pydantic==2.10.6
PyYAML==6.0.2
